Update version, add new method to return meta string
New method lets other classes grab the meta name so that if
it is updated in the future, it will continue to work.

Rename some of the params in "manage_users_column_content" to match
those given in the wordpress docs:
- https://developer.wordpress.org/reference/hooks/manage_users_custom_column/

Currently, downloading the plugin from wpackagist doesn't return
the version of the plugin with the return statement in manage_users_column_content.
Hoping that updating the version will fix that.